pipeline/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "msp-lead-magnet"
version = "0.0.1"
description = "MSP lead magnet pipeline: scrape, enrich, score, output."
requires-python = ">=3.11"
dependencies = [
    "python-jobspy>=1.1.70",
    "feedparser>=6.0",
    "requests>=2.31",
    "beautifulsoup4>=4.12",
    "rapidfuzz>=3.0",
    "anthropic>=0.40",
    "google-genai>=0.3",
    "python-dotenv>=1.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.6",
    "mypy>=1.10",
]

[tool.setuptools]
packages = ["msp_pipeline", "msp_pipeline.sources"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_unused_ignores = true
warn_redundant_casts = true

[[tool.mypy.overrides]]
module = ["jobspy", "feedparser", "requests", "requests.*"]
ignore_missing_imports = true
